Rename testing workflow and Docker images
Rename the testing workflow to `Build and push custom Docker images`, which is
more appropriate since it doesn't imply a use case; especially if users fork
this repository and want to use that workflow to build custom images.
Also rename the resulting Docker images to be suffixed with `-custom` instead
of `-testing`.